requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alibabacloud-cdn20180510==3.2.0
alibabacloud-credentials==0.3.6
alibabacloud-gateway-spi==0.0.2
alibabacloud-ram20150501==1.1.0
alibabacloud-resourcedirectorymaster20220419==2.1.0
alibabacloud-sts20150401==1.1.4
alibabacloud-tea==0.4.0
alibabacloud-tea-openapi==0.3.12
alibabacloud-tea-util==0.3.13
alibabacloud_endpoint_util==0.0.3
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_xml==0.0.2
aliyun-python-sdk-cdn==3.8.8
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
aniso8601==9.0.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
boto3==1.28.62
botocore==1.31.85
cachelib==0.9.0
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.12
circuitbreaker==1.4.0
click==8.1.7
colorama==0.4.6
cos-python-sdk-v5==1.9.26
crcmod==1.7
cryptography==41.0.5
DateTime==5.4
Deprecated==1.2.14
Flask==3.0.0
Flask-APScheduler==1.13.1
Flask-Caching==2.1.0
Flask-Limiter==3.3.1
Flask-RESTful==0.3.10
frozenlist==1.4.1
idna==2.7
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==0.10.0
limits==3.6.0
loguru==0.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.5
numpy==2.0.2
oci==2.115.0
ordered-set==4.1.0
oss2==2.19.1
packaging==23.2
pandas==2.2.3
prettytable==3.9.0
prometheus-client==0.17.0
pycparser==2.21
pycryptodome==3.19.0
Pygments==2.16.1
PyMySQL==1.1.0
pyOpenSSL==23.3.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.27.1
rich==13.6.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.8
s3transfer==0.7.0
six==1.16.0
tccli==3.0.1223.1
tencentcloud-sdk-python==3.0.1274
tencentcloud-sdk-python-common==3.0.1055
tencentcloud-sdk-python-cvm==3.0.1055
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2024.2
tzlocal==5.2
urllib3==1.26.20
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.15.0
xmltodict==0.13.0
yarl==1.9.4
zope.interface==6.1
